Outsource redis getting and empty redis DB class
